Fix #5

+++ src/nn-builder/src/arch/initializers.cc
@@ -5,7 +5,7 @@
 namespace nn {
 namespace arch {
 
-std::vector<wasmpp::DataEntry> XavierDistribution(uint32_t size, uint32_t n_in, uint32_t n_out, bool uniform, uint64_t seed) {
+std::vector<wasmpp::DataEntry> XavierDistribution(uint32_t size, uint32_t n_in, uint32_t n_out, bool uniform, uint32_t seed) {
   if(uniform) {
     float limit = sqrtf(6.0f / (n_in + n_out));
     return UniformDistribution(size, -limit, limit, seed);
@@ -16,7 +16,7 @@ std::vector<wasmpp::DataEntry> XavierDistribution(uint32_t size, uint32_t n_in, 
   return GaussianDistribution(size, mean, std_dev, seed);
 }
 
-std::vector<wasmpp::DataEntry> LeCunDistribution(uint32_t size, uint32_t n_in, bool uniform, uint64_t seed) {
+std::vector<wasmpp::DataEntry> LeCunDistribution(uint32_t size, uint32_t n_in, bool uniform, uint32_t seed) {
   if(uniform) {
     float limit = sqrtf(3.0f / n_in);
     return UniformDistribution(size, -limit, limit, seed);
@@ -27,7 +27,7 @@ std::vector<wasmpp::DataEntry> LeCunDistribution(uint32_t size, uint32_t n_in, b
   return GaussianDistribution(size, mean, std_dev, seed);
 }
 
-std::vector<wasmpp::DataEntry> GaussianDistribution(uint32_t size, float mean, float std_dev, uint64_t seed) {
+std::vector<wasmpp::DataEntry> GaussianDistribution(uint32_t size, float mean, float std_dev, uint32_t seed) {
   std::default_random_engine generator(seed);
   std::vector<wasmpp::DataEntry> entries;
   std::normal_distribution<float> distribution(mean, std_dev);
@@ -37,7 +37,7 @@ std::vector<wasmpp::DataEntry> GaussianDistribution(uint32_t size, float mean, f
   return entries;
 }
 
-std::vector<wasmpp::DataEntry> UniformDistribution(uint32_t size, float low, float high, uint64_t seed) {
+std::vector<wasmpp::DataEntry> UniformDistribution(uint32_t size, float low, float high, uint32_t seed) {
   std::default_random_engine generator(seed);
   std::vector<wasmpp::DataEntry> entries;
   std::uniform_real_distribution<float> distribution(low, high);
